Add snapshot tests for mondrian-charts-rs + use f64 math

+++ mondrian-charts/src/constants.rs
@@ -1,2 +1,2 @@
-pub const BAR_PADDING: f32 = 0.2;
-pub const BAR_PLUS_PADDING: f32 = 1. + BAR_PADDING;
+pub const BAR_PADDING: f64 = 0.2;
+pub const BAR_PLUS_PADDING: f64 = 1. + BAR_PADDING;
