+++ mondrian-charts/src/utils/attach_suggestions_to_x_axis.rs
@@ -6,7 +6,7 @@ use crate::types::{Axis, Buckets};
 pub(crate) fn attach_suggestions_to_x_axis<T>(
     x_axis: &mut Axis,
     buckets: &Buckets<T>,
-    interval: f32,
+    interval: f64,
 ) {
     if interval <= 0. {
         return;
@@ -29,7 +29,7 @@ pub(crate) fn attach_suggestions_to_x_axis<T>(
     x_axis.tick_suggestions = Some(suggestions);
 }
 
-fn get_first_bucket_time<T>(buckets: &Buckets<T>) -> Option<f32> {
+fn get_first_bucket_time<T>(buckets: &Buckets<T>) -> Option<f64> {
     let mut first_bucket_timestamp = None;
     for timestamp in buckets.keys() {
         if first_bucket_timestamp
